requirements.txt
# ============================================================================
# AI-Native MVP - Python Dependencies (Fase 8)
# ============================================================================

# Core Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.1.1

# Database
sqlalchemy==2.0.37
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis Cache
redis==5.2.1
hiredis==3.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.2.1

# HTTP Client
httpx>=0.27.0,<0.28.0

# Monitoring & Metrics
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0

# LLM Providers
openai==1.59.8
anthropic==0.42.0
google-generativeai==0.8.3
mistralai==1.2.4

# LangChain & RAG
langchain==0.3.17
langchain-community==0.3.16
langchain-openai==0.2.14
langchain-anthropic==0.3.5
langchain-mistralai==0.2.2
langgraph==0.2.60
langgraph-checkpoint-sqlite==2.0.8
chromadb==0.5.23
sentence-transformers==3.3.1

# Ollama (Local LLM)
ollama==0.4.6

# Document Processing
pypdf==5.1.0
PyPDF2==3.0.1
pymupdf==1.25.2
python-docx==1.1.2
markdown==3.7

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
pyyaml==6.0.2
colorama==0.4.6
json-repair==0.32.0

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Type Checking & Linting
mypy==1.14.1
ruff==0.8.6

# Date & Time
python-dateutil==2.9.0.post0
